refactor(redux): split taskSlice into uiSlice + formSlice (Phase 1)
taskSlice (431 lines, 4 concerns) split into three focused slices:
- uiSlice: isCreate, isTaskDetail, isPopup, popupMode, isHover,
  isSidebarPinned, activeMenu, selectedTask + selectedTask sync
  extraReducers for updatedTask/completedTask fulfilled
- formSlice: formTask, progress + auto-reset extraReducer on
  createNewTask.fulfilled
- taskSlice: server state only (~170 lines)

All 21 import sites updated; TaskForm and GroupTaskForm (missed by
initial grep) fixed — both were silently reading isCreate/selectedTask
from state.tasks after the move. Also fixes a pre-existing isSideBarPinned
typo in GroupTaskForm that kept the clear-all button always visible.

Tests: 109 passing (17 files), 0 lint errors.
