+++ frontend/src/__tests__/redux/uiSlice.test.js
@@ -1,0 +1,130 @@
+import { describe, it, expect } from "vitest";
+import reducer, {
+  toggleCreatePopup,
+  setActiveMenu,
+  togglePopup,
+  setHover,
+  toggleSidebarPinned,
+  setSelectedTask,
+} from "../../redux/uiSlice";
+
+const init = () => reducer(undefined, { type: "@@INIT" });
+
+describe("uiSlice — initial state", () => {
+  it("starts with all popups closed and no selection", () => {
+    const state = init();
+    expect(state.isCreate).toBe(false);
+    expect(state.isTaskDetail).toBe(false);
+    expect(state.isPopup).toBe(false);
+    expect(state.popupMode).toBe("");
+    expect(state.isHover).toBe(null);
+    expect(state.isSidebarPinned).toBe(false);
+    expect(state.activeMenu).toBe("");
+    expect(state.selectedTask).toBe(null);
+  });
+});
+
+describe("uiSlice — UI toggles", () => {
+  it("setActiveMenu stores the menu name", () => {
+    expect(reducer(init(), setActiveMenu("home")).activeMenu).toBe("home");
+  });
+
+  it("setHover stores the hovered id", () => {
+    expect(reducer(init(), setHover("t1")).isHover).toBe("t1");
+  });
+
+  it("toggleCreatePopup flips isCreate", () => {
+    expect(reducer(init(), toggleCreatePopup()).isCreate).toBe(true);
+  });
+
+  it("togglePopup flips isPopup and records mode", () => {
+    const state = reducer(init(), togglePopup("delete"));
+    expect(state.isPopup).toBe(true);
+    expect(state.popupMode).toBe("delete");
+  });
+
+  it("togglePopup with no arg sets popupMode to empty string", () => {
+    const open = reducer(init(), togglePopup("delete"));
+    const closed = reducer(open, togglePopup(""));
+    expect(closed.isPopup).toBe(false);
+    expect(closed.popupMode).toBe("");
+  });
+
+  it("toggleSidebarPinned flips the flag", () => {
+    expect(reducer(init(), toggleSidebarPinned()).isSidebarPinned).toBe(true);
+  });
+
+  it("setSelectedTask stores the selected task and opens detail", () => {
+    const task = { _id: "a", title: "X" };
+    const state = reducer(init(), setSelectedTask(task));
+    expect(state.selectedTask).toEqual(task);
+    expect(state.isTaskDetail).toBe(true);
+  });
+
+  it("setSelectedTask(null) clears the task and closes detail", () => {
+    const open = { ...init(), selectedTask: { _id: "a" }, isTaskDetail: true };
+    const state = reducer(open, setSelectedTask(null));
+    expect(state.selectedTask).toBe(null);
+    expect(state.isTaskDetail).toBe(false);
+  });
+
+  it("selecting a second task keeps detail open (no toggle bug)", () => {
+    const open = reducer(init(), setSelectedTask({ _id: "a" }));
+    const next = reducer(open, setSelectedTask({ _id: "b" }));
+    expect(next.selectedTask._id).toBe("b");
+    expect(next.isTaskDetail).toBe(true);
+  });
+});
+
+describe("uiSlice — extraReducers: selectedTask sync", () => {
+  it("updatedTask.fulfilled updates selectedTask when IDs match", () => {
+    const state = {
+      ...init(),
+      selectedTask: { _id: "a", title: "old", category: "x" },
+      isTaskDetail: true,
+    };
+    const next = reducer(state, {
+      type: "task/updateTask/fulfilled",
+      payload: { _id: "a", title: "new", category: "y" },
+    });
+    expect(next.selectedTask.title).toBe("new");
+    expect(next.selectedTask.category).toBe("y");
+  });
+
+  it("updatedTask.fulfilled does not touch selectedTask when IDs differ", () => {
+    const state = {
+      ...init(),
+      selectedTask: { _id: "b", title: "other" },
+    };
+    const next = reducer(state, {
+      type: "task/updateTask/fulfilled",
+      payload: { _id: "a", title: "new" },
+    });
+    expect(next.selectedTask._id).toBe("b");
+  });
+
+  it("completedTask.fulfilled updates selectedTask status when IDs match", () => {
+    const state = {
+      ...init(),
+      selectedTask: { _id: "a", status: "pending" },
+      isTaskDetail: true,
+    };
+    const next = reducer(state, {
+      type: "task/completedTask/fulfilled",
+      payload: { _id: "a", status: "completed" },
+    });
+    expect(next.selectedTask.status).toBe("completed");
+  });
+
+  it("completedTask.fulfilled does not touch selectedTask when IDs differ", () => {
+    const state = {
+      ...init(),
+      selectedTask: { _id: "b", status: "pending" },
+    };
+    const next = reducer(state, {
+      type: "task/completedTask/fulfilled",
+      payload: { _id: "a", status: "completed" },
+    });
+    expect(next.selectedTask.status).toBe("pending");
+  });
+});
